fixed some issues with injectFlip

+++ src/vm/alu.cpp
@@ -54,7 +54,7 @@ static std::string decode_fclass(uint16_t res) {
     }
     case AluOp::kInjectFlip: {
     // a = x2 (source and destination), b = x0 (ignored)
-    int32_t value = static_cast<int32_t>(a);  // Extract lower 32 bits as signed int
+    uint64_t value = a;  // Extract lower 32 bits as signed int
 
     // Seed with random device + time for true randomness 🌱
     static std::random_device rd;
@@ -71,12 +71,12 @@ static std::string decode_fclass(uint16_t res) {
       int pos = bit_pos(gen);
 
       // FLIP THAT BIT! 🎭
-      value ^= (1 << pos);
+      value ^= (1ULL << pos);
       did_flip = true;
     }
 
     // Return new value in x2 (zero-extended to 64-bit), and whether flip occurred
-    return {static_cast<uint64_t>(static_cast<uint32_t>(value)), did_flip};
+    return {value, did_flip};
   }
     case AluOp::kAddw: {
       auto sa = static_cast<int32_t>(a);
